pyproject.toml
[tool.poetry]
name = "python-scripts"
version = "0.1.0"
description = "Collection of persnal scripts written in Python by Daniel Gill"
authors = ["Daniel Gill <dwgill@outlook.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aiosqlite = "^0.20.0"
obsidiantools = "^0.10.0"
typer = "^0.9.0"
python-frontmatter = "^1.1.0"
python-dotenv = "^1.0.1"
cryptography = "^42.0.5"
pyairtable = "^2.3.3"
pydantic = "^2.6.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
